🎨 Refactor: 重构 AICardGenerationService 以应用策略模式
将原来庞大的 AICardGenerationService 类基于策略模式 (Strategy Pattern) 进行了拆分：
- 提取 `CardGenerationStrategy` 抽象接口。
- 分离 AI 卡片生成逻辑到 `AiCardGenerationStrategy`。
- 分离本地备用模板卡片生成逻辑到 `FallbackCardGenerationStrategy`。
- 将纯静态的多语言处理、日期处理等方法转移至公共工具类 `CardGenerationUtils`。
- 将用于后台 Isolate 运行的 SVG 处理方法独立到 `svg_processing_isolate.dart`，以保证 Isolate 通信的稳定性并实现代码解耦。
- `AICardGenerationService` 仅保留对外统一 API 暴露及公共图片保存逻辑。
