+++ lib/services/ai_card_generation_service.dart
@@ -1,37 +1,29 @@
 import 'package:flutter/foundation.dart';
 import 'dart:io';
-import 'package:uuid/uuid.dart';
+import 'package:flutter/widgets.dart';
+
 // 条件导入：Web平台使用stub实现，其他平台使用gal
 import '../utils/stub_implementations.dart'
     if (dart.library.io) 'package:gal/gal.dart' as gal;
 import '../models/quote_model.dart';
 import '../models/generated_card.dart';
-import '../constants/ai_card_prompts.dart';
-import '../constants/card_templates.dart';
 import '../utils/app_logger.dart';
-import '../utils/string_utils.dart';
 import 'ai_service.dart';
 import 'settings_service.dart';
-import 'database_service.dart';
 import 'svg_to_image_service.dart';
-import 'package:flutter/widgets.dart'; // 为新增的BuildContext参数添加导入
+
+import 'ai_card_generation_strategies/ai_card_generation_strategy.dart';
+import 'ai_card_generation_strategies/fallback_card_generation_strategy.dart';
 
 /// AI卡片生成服务
-// TODO(refactor): This file exceeds 1500 lines. Consider splitting card generation
-// logic into separate strategies for different card types or layouts.
 class AICardGenerationService {
-  final AIService _aiService;
   final SettingsService _settingsService;
+  final AiCardGenerationStrategy _aiStrategy;
+  final FallbackCardGenerationStrategy _fallbackStrategy;
 
-  static final RegExp _viewBoxRegex = RegExp(r'viewBox="([^"]+)"');
-  static final RegExp _splitSpaceCommaRegex = RegExp(r'[\s,]+');
-  static final RegExp _widthRegex = RegExp(r'width="([^"]+)"');
-  static final RegExp _heightRegex = RegExp(r'height="([^"]+)"');
-  static final RegExp _rectFallbackRegex =
-      RegExp(r'<rect[^>]*width="([^"]+)"[^>]*height="([^"]+)"');
-  static final RegExp _numericRegex = RegExp('[^0-9.-]');
-
-  AICardGenerationService(this._aiService, this._settingsService);
+  AICardGenerationService(AIService aiService, this._settingsService)
+      : _aiStrategy = AiCardGenerationStrategy(aiService, _settingsService),
+        _fallbackStrategy = FallbackCardGenerationStrategy();
 
   /// 获取当前语言代码 (zh, en, ja, fr, etc.)
   String get _currentLanguageCode {
@@ -63,379 +55,44 @@ class AICardGenerationService {
     bool isRegeneration = false,
     CardType? excludeType,
   }) async {
+    final languageCode = _currentLanguageCode;
+
     // 如果用户关闭了 AI 生成功能，则直接使用模板（功能仍可用，只是没有AI增强）
     if (!isEnabled) {
       AppLogger.i('AI卡片生成已关闭，使用本地模板生成', source: 'AICardGeneration');
-      return _buildFallbackCard(note,
-          isRegeneration: isRegeneration,
-          brandName: brandName,
-          excludeType: excludeType);
+      return _fallbackStrategy.generate(
+        note: note,
+        brandName: brandName,
+        languageCode: languageCode,
+        customStyle: customStyle,
+        isRegeneration: isRegeneration,
+        excludeType: excludeType,
+      );
     }
 
-    // 根据用户语言设置决定卡片元数据语言
-    final languageCode = _currentLanguageCode;
-
     try {
-      // 1. 智能选择最适合的提示词
-      var prompt = _selectBestPrompt(note, customStyle,
-          brandName: brandName,
-          isRegeneration: isRegeneration,
-          languageCode: languageCode);
-
-      // 1.1 根据用户语言设置追加语言统一指令
-      String langDirective;
-      switch (languageCode) {
-        case 'zh':
-          langDirective =
-              '使用全中文作为所有底部元数据（日期、天气、时间段等），不要出现英文单词（例如用"雨""晨间""夜晚"而不是 rain/Morning）。如果某项信息缺失可以省略，不要编造。';
-          break;
-        case 'ja':
-          langDirective =
-              'Use Japanese for all footer metadata (date, weather, period). Do not mix English. If some info is missing, omit it.';
-          break;
-        case 'fr':
-          langDirective =
-              'Use French for all footer metadata (date, weather, period). Do not mix English. If some info is missing, omit it.';
-          break;
-        case 'en':
-        default:
-          langDirective =
-              'Use English for all footer metadata (date, weather, period). Do not mix Chinese. If some info is missing, omit it.';
-          break;
-      }
-
-      prompt = '$prompt\n\n### 语言 / Language Constraint\n$langDirective';
-
-      AppLogger.i(
-        '开始AI生成SVG卡片，内容长度: ${note.content.length}',
-        source: 'AICardGeneration',
-      );
-
-      // 2. 调用AI生成SVG
-      final svgContent = await _generateSVGContent(prompt);
-
-      // 3. 在后台 isolate 中处理 SVG (清理 + 元数据补全)
-      final processingData = AICardProcessingData(
-        svgContent: svgContent,
+      return await _aiStrategy.generate(
+        note: note,
         brandName: brandName,
-        date: _formatDate(note.date, languageCode: languageCode),
-        location: note.location,
-        weather: note.weather,
-        temperature: note.temperature,
-        author: note.sourceAuthor,
-        source: note.fullSource,
-        dayPeriod: note.dayPeriod,
         languageCode: languageCode,
-      );
-
-      final result = await compute(processSVGTask, processingData);
-
-      // 4. 重放日志
-      for (final log in result.logs) {
-        switch (log.level) {
-          case 'ERROR':
-            AppLogger.e(log.message, source: 'AICardGeneration');
-            break;
-          case 'WARN':
-            AppLogger.w(log.message, source: 'AICardGeneration');
-            break;
-          case 'INFO':
-            AppLogger.i(log.message, source: 'AICardGeneration');
-            break;
-          case 'DEBUG':
-          default:
-            AppLogger.d(log.message, source: 'AICardGeneration');
-            break;
-        }
-      }
-
-      final cleanedSVG = result.svg;
-
-      AppLogger.i(
-        'AI生成SVG成功，长度: ${cleanedSVG.length}',
-        source: 'AICardGeneration',
-      );
-
-      // 5. 创建卡片对象（AI生成默认类型：knowledge）
-      return GeneratedCard(
-        id: const Uuid().v4(),
-        noteId: note.id!,
-        originalContent: StringUtils.removeObjectReplacementChar(note.content),
-        svgContent: cleanedSVG,
-        type: CardType.knowledge,
-        createdAt: DateTime.now(),
-        author: note.sourceAuthor,
-        source: note.fullSource,
-        location: note.location,
-        weather: note.weather,
-        temperature: note.temperature,
-        date: note.date,
-        dayPeriod: note.dayPeriod,
+        customStyle: customStyle,
+        isRegeneration: isRegeneration,
+        excludeType: excludeType,
       );
     } catch (e) {
       AppLogger.w('AI生成失败，使用回退模板: $e', source: 'AICardGeneration');
-      return _buildFallbackCard(note,
-          isRegeneration: isRegeneration, brandName: brandName);
+      return _fallbackStrategy.generate(
+        note: note,
+        brandName: brandName,
+        languageCode: languageCode,
+        customStyle: customStyle,
+        isRegeneration: isRegeneration,
+        excludeType: excludeType,
+      );
     }
   }
 
-  /// 本地模板回退封装(AI关闭或失败时使用)
-  GeneratedCard _buildFallbackCard(Quote note,
-      {bool isRegeneration = false,
-      required String brandName,
-      CardType? excludeType}) {
-    // 智能检测最适合的模板类型
-    final cardType = _determineTemplateType(note,
-        isRegeneration: isRegeneration, excludeType: excludeType);
-    final languageCode = _currentLanguageCode;
-    final cleanContent = StringUtils.removeObjectReplacementChar(note.content);
-    final fallbackSVG = CardTemplates.getTemplateByType(
-      brandName: brandName,
-      type: cardType,
-      content: cleanContent,
-      author: note.sourceAuthor,
-      date: _formatDate(note.date, languageCode: languageCode),
-      source: note.fullSource,
-      location: note.location,
-      weather: _localizeWeather(note.weather, languageCode: languageCode),
-      temperature: note.temperature,
-      dayPeriod: _localizeDayPeriod(note.dayPeriod, languageCode: languageCode),
-    );
-
-    return GeneratedCard(
-      id: const Uuid().v4(),
-      noteId: note.id!,
-      originalContent: cleanContent,
-      svgContent: fallbackSVG,
-      type: cardType,
-      createdAt: DateTime.now(),
-      author: note.sourceAuthor,
-      source: note.fullSource,
-      location: note.location,
-      weather: note.weather,
-      temperature: note.temperature,
-      date: note.date,
-      dayPeriod: note.dayPeriod,
-    );
-  }
-
-  /// 智能决定模板类型（基于标签、内容和随机性）
-  CardType _determineTemplateType(Quote note,
-      {bool isRegeneration = false, CardType? excludeType}) {
-    // 重新生成时完全随机，跳过内容匹配
-    if (isRegeneration) {
-      final allTypes = CardType.values.where((t) => t != excludeType).toList();
-      if (allTypes.isEmpty) return CardType.knowledge;
-      final randomcheck = DateTime.now().microsecondsSinceEpoch;
-      return allTypes[randomcheck % allTypes.length];
-    }
-
-    // 1. 优先匹配 Hitokoto 官方分类 (如果 categoryId 匹配)
-    if (note.categoryId != null) {
-      switch (note.categoryId) {
-        case DatabaseService.defaultCategoryIdAnime: // 动画 -> 几何/视觉
-          return CardType.geometric;
-        case DatabaseService.defaultCategoryIdComic: // 漫画 -> 几何/视觉
-          return CardType.geometric;
-        case DatabaseService.defaultCategoryIdGame: // 游戏 -> 赛博/科技
-          return CardType.cyberpunk;
-        case DatabaseService.defaultCategoryIdNovel: // 小说 -> 复古/纸张
-          return CardType.retro;
-        case DatabaseService.defaultCategoryIdPoem: // 诗词 -> 水墨/禅意
-          return CardType.ink;
-        case DatabaseService.defaultCategoryIdPhilosophy: // 哲学 -> 哲学/深邃
-          return CardType.philosophical;
-        case DatabaseService.defaultCategoryIdOriginal: // 原创 -> 情感/日记
-          return CardType.emotional;
-        case DatabaseService.defaultCategoryIdMusic: // 音乐 -> 情感/日记
-          return CardType.emotional;
-        case DatabaseService
-              .defaultCategoryIdInternet: // 网络 -> 开发者/代码 (通常是网络段子或技术梗)
-          return CardType.dev;
-        case DatabaseService.defaultCategoryIdMovie: // 影视 -> 引用/剧照感
-          return CardType.quote;
-        case DatabaseService.defaultCategoryIdJoke: // 抖机灵 -> 极简/留白 (突出笑点)
-          return CardType.minimalist;
-      }
-    }
-
-    final content = note.content.toLowerCase();
-    final keywords = note.keywords?.map((e) => e.toLowerCase()).toList() ?? [];
-
-    // 2. 匹配明确的关键词/标签 (支持中英文)
-    if (_hasKeyword(content, keywords, [
-      '代码',
-      '编程',
-      '开发',
-      'code',
-      'dev',
-      'programming',
-      'bug',
-      'flutter',
-      'dart',
-      'api'
-    ])) {
-      return CardType.dev;
-    }
-
-    if (_hasKeyword(content, keywords, [
-      '日记',
-      '心情',
-      '感受',
-      'diary',
-      'mood',
-      'feeling',
-      'emotion',
-      'love',
-      '悲伤',
-      '快乐'
-    ])) {
-      return CardType.emotional;
-    }
-
-    if (_hasKeyword(content, keywords, [
-      '学习',
-      '笔记',
-      '复习',
-      'study',
-      'note',
-      'learn',
-      'exam',
-      'research',
-      'paper',
-      '学术'
-    ])) {
-      return CardType.academic;
-    }
-
-    if (_hasKeyword(content, keywords, [
-      '自然',
-      '风景',
-      'nature',
-      'tree',
-      'flower',
-      'mountain',
-      'river',
-      'green',
-      'eco'
-    ])) {
-      return CardType.nature;
-    }
-
-    if (_hasKeyword(content, keywords,
-        ['思考', '哲学', '意义', 'philosophy', 'think', 'mind', 'reason', 'truth'])) {
-      return CardType.philosophical;
-    }
-
-    if (_hasKeyword(content, keywords, [
-      '历史',
-      '复古',
-      '旧',
-      'retro',
-      'history',
-      'old',
-      'vintage',
-      'memory',
-      'time'
-    ])) {
-      return CardType.retro;
-    }
-
-    if (_hasKeyword(content, keywords, [
-      '禅',
-      '道',
-      '静',
-      'ink',
-      'zen',
-      'tao',
-      'chinese',
-      'calligraphy',
-      'buddha'
-    ])) {
-      return CardType.ink;
-    }
-
-    if (_hasKeyword(content, keywords, [
-      '赛博',
-      '未来',
-      '科技',
-      'cyber',
-      'future',
-      'tech',
-      'neon',
-      'glitch',
-      'punk'
-    ])) {
-      return CardType.cyberpunk;
-    }
-
-    if (_hasKeyword(content, keywords,
-        ['几何', '设计', '艺术', 'geo', 'design', 'art', 'shape', 'abstract'])) {
-      return CardType.geometric;
-    }
-
-    // 2. 基于元数据的启发式规则
-    final hasAuthor =
-        note.sourceAuthor != null && note.sourceAuthor!.isNotEmpty;
-    // 如果是短文本且有作者，很可能是名言
-    if (hasAuthor && note.content.length < 100) {
-      // 50% 概率使用专门的引用模板
-      if (DateTime.now().millisecond % 2 == 0) {
-        return CardType.quote;
-      }
-    }
-
-    // 3. 随机回退
-    final allTypes = [
-      CardType.knowledge,
-      CardType.quote,
-      CardType.philosophical,
-      CardType.minimalist,
-      CardType.nature,
-      CardType.retro,
-      CardType.ink,
-      CardType.cyberpunk,
-      CardType.geometric,
-      CardType.academic,
-      CardType.emotional,
-      CardType.dev,
-      CardType.mindful,
-      CardType.neonCyber,
-      CardType.classicSerif,
-      CardType.modernPop,
-      CardType.softGradient,
-      CardType.polaroid,
-      CardType.magazine,
-      CardType.sotaModern,
-    ].where((t) => t != excludeType).toList();
-
-    if (allTypes.isEmpty) return CardType.sotaModern;
-
-    final randomcheck = DateTime.now().microsecondsSinceEpoch;
-    // 增加 SOTA Modern 的出现概率 (20%)
-    if (randomcheck % 5 == 0 && excludeType != CardType.sotaModern) {
-      return CardType.sotaModern;
-    }
-    return allTypes[randomcheck % allTypes.length];
-  }
-
-  /// 检查内容或关键词中是否包含目标词汇
-  bool _hasKeyword(
-      String content, List<String> keywords, List<String> targets) {
-    // 检查内容
-    for (final target in targets) {
-      if (content.contains(target)) return true;
-    }
-    // 检查关键词
-    for (final keyword in keywords) {
-      for (final target in targets) {
-        if (keyword.contains(target)) return true;
-      }
-    }
-    return false;
-  }
-
-  /// 批量生成（用于周期报告）
+  /// 批量为多条笔记生成卡片（主要用于精选内容展示）
   Future<List<GeneratedCard>> generateFeaturedCards({
     required String brandName,
     required List<Quote> notes,
@@ -477,7 +134,7 @@ class AICardGenerationService {
 
         // 如果失败卡片太多，停止生成
         if (errors.length > maxCards ~/ 2) {
-          logError('批量生成失败率过高，停止生成', source: 'AICardGenerationService');
+          AppLogger.e('批量生成失败率过高，停止生成', source: 'AICardGenerationService');
           break;
         }
 
@@ -486,7 +143,7 @@ class AICardGenerationService {
     }
 
     if (cards.isEmpty && errors.isNotEmpty) {
-      throw AICardGenerationException('批量生成完全失败: ${errors.join('; ')}');
+      throw Exception('批量生成完全失败: ${errors.join('; ')}');
     }
 
     AppLogger.i(
@@ -497,157 +154,7 @@ class AICardGenerationService {
     return cards;
   }
 
-  /// 生成SVG内容
-  Future<String> _generateSVGContent(String prompt) async {
-    // 检查AI服务是否可用
-    if (!await _aiService.hasValidApiKeyAsync()) {
-      throw const AICardGenerationException('请先在设置中配置 API Key');
-    }
-
-    try {
-      // 获取当前provider
-      final multiSettings = _settingsService.multiAISettings;
-      final currentProvider = multiSettings.currentProvider;
-
-      if (currentProvider == null) {
-        throw const AICardGenerationException('未找到可用的AI提供商配置');
-      }
-
-      // 使用专门的SVG生成方法
-      final svgContent = await _aiService.generateSVG(prompt);
-
-      // 验证生成的SVG是否有效
-      if (svgContent.trim().isEmpty) {
-        throw const AICardGenerationException('AI返回了空的SVG内容');
-      }
-
-      return svgContent;
-    } catch (e) {
-      AppLogger.e('AI SVG生成失败: $e', error: e, source: 'AICardGeneration');
-      rethrow; // 重新抛出异常，让上层处理回退
-    }
-  }
-
-  /// Isolate worker function responsible for cleaning SVG and injecting metadata.
-  ///
-  /// This runs in a background isolate to prevent UI jank.
-  static Future<AICardProcessingResult> processSVGTask(
-      AICardProcessingData data) async {
-    final logs = <AICardProcessingLog>[];
-    logs.add(AICardProcessingLog(
-        'DEBUG', '开始清理SVG内容，原始长度: ${data.svgContent.length}'));
-
-    try {
-      // 1. Clean SVG
-      String cleaned = _cleanSVGContentStatic(data.svgContent, logs);
-
-      // 2. Ensure Metadata
-      cleaned = _ensureMetadataPresenceStatic(
-        cleaned,
-        brandName: data.brandName,
-        date: data.date,
-        location: data.location,
-        weather: data.weather,
-        temperature: data.temperature,
-        author: data.author,
-        source: data.source,
-        dayPeriod: data.dayPeriod,
-        languageCode: data.languageCode,
-        logs: logs,
-      );
-
-      logs.add(AICardProcessingLog('DEBUG', 'SVG处理完成，最终长度: ${cleaned.length}'));
-      return AICardProcessingResult(svg: cleaned, logs: logs);
-    } catch (e) {
-      logs.add(AICardProcessingLog('ERROR', 'SVG处理失败: $e'));
-      rethrow;
-    }
-  }
-
-  /// 静态清理SVG内容 (用于Isolate)
-  static String _cleanSVGContentStatic(
-      String response, List<AICardProcessingLog> logs) {
-    String cleaned = response.trim();
-
-    // 移除常见的markdown标记和说明文字
-    cleaned = cleaned
-        .replaceAll('```svg', '')
-        .replaceAll('```xml', '')
-        .replaceAll('```html', '')
-        .replaceAll('```', '')
-        .replaceAll('`', '')
-        .trim();
-
-    // 移除可能的说明文字（在SVG前后）
-    final lines = cleaned.split('\n');
-    final svgLines = <String>[];
-    bool inSvg = false;
-    bool foundSvgStart = false;
-
-    for (final line in lines) {
-      final trimmedLine = line.trim();
-
-      // 跳过空行和注释行（除非在SVG内部）
-      if (!inSvg && (trimmedLine.isEmpty || trimmedLine.startsWith('//'))) {
-        continue;
-      }
-
-      // 检测SVG开始
-      if (trimmedLine.startsWith('<svg')) {
-        inSvg = true;
-        foundSvgStart = true;
-        svgLines.add(line);
-        continue;
-      }
-
-      // 检测SVG结束
-      if (trimmedLine.contains('</svg>')) {
-        svgLines.add(line);
-        break; // SVG结束后停止处理
-      }
-
-      // 在SVG内部时保留所有行
-      if (inSvg) {
-        svgLines.add(line);
-      }
-    }
-
-    cleaned = svgLines.join('\n').trim();
-
-    // 如果没有找到完整的SVG，尝试简单的字符串提取
-    if (!foundSvgStart ||
-        !cleaned.contains('<svg') ||
-        !cleaned.contains('</svg>')) {
-      logs.add(AICardProcessingLog('WARN', '未找到完整SVG，尝试字符串提取...'));
-
-      final svgStartIndex = response.indexOf('<svg');
-      if (svgStartIndex >= 0) {
-        final svgEndIndex = response.lastIndexOf('</svg>');
-        if (svgEndIndex > svgStartIndex) {
-          cleaned = response.substring(svgStartIndex, svgEndIndex + 6);
-          logs.add(
-              AICardProcessingLog('INFO', '字符串提取成功，SVG长度: ${cleaned.length}'));
-        }
-      }
-    }
-
-    // 验证SVG基本结构
-    if (!_isValidSVGStructure(cleaned)) {
-      throw const AICardGenerationException('AI返回的内容不包含有效的SVG代码');
-    }
-
-    // 标准化SVG属性
-    cleaned = _normalizeSVGAttributes(cleaned);
-
-    // 验证SVG内容安全性
-    if (!_isSafeSVGContent(cleaned, logs)) {
-      throw const AICardGenerationException('SVG内容包含不安全的元素');
-    }
-
-    return cleaned;
-  }
-
-  /// 保存卡片为图片
+  /// 将生成的卡片保存为图片（带权限检查）
   Future<String> saveCardAsImage(
     GeneratedCard card, {
     int width = 400,
@@ -798,748 +305,9 @@ class AICardGenerationService {
     return savedFiles;
   }
 
-  /// 格式化日期 - 支持多语言
-  /// [languageCode] 语言代码 ('zh', 'en', 'ja', 'fr', etc.)
-  String _formatDate(String dateStr, {String languageCode = 'zh'}) {
-    try {
-      final date = DateTime.parse(dateStr);
-      switch (languageCode) {
-        case 'zh':
-        case 'ja':
-          // 中日文: 2026年1月22日
-          return '${date.year}年${date.month}月${date.day}日';
-        case 'fr':
-          // 法文: 22/01/2026
-          final d = date.day.toString().padLeft(2, '0');
-          final m = date.month.toString().padLeft(2, '0');
-          return '$d/$m/${date.year}';
-        case 'en':
-        default:
-          // 英文格式: Jan 22, 2026
-          const months = [
-            'Jan',
-            'Feb',
-            'Mar',
-            'Apr',
-            'May',
-            'Jun',
-            'Jul',
-            'Aug',
-            'Sep',
-            'Oct',
-            'Nov',
-            'Dec'
-          ];
-          return '${months[date.month - 1]} ${date.day}, ${date.year}';
-      }
-    } catch (e) {
-      return dateStr;
-    }
-  }
-
   /// 检查AI卡片生成功能是否启用
   bool get isEnabled {
     // 从设置服务中获取AI卡片生成开关状态
     return _settingsService.aiCardGenerationEnabled;
   }
-
-  /// 智能选择最适合的提示词（改进：增加随机性和变化）
-  String _selectBestPrompt(Quote note, String? customStyle,
-      {required String brandName,
-      bool isRegeneration = false,
-      String languageCode = 'zh'}) {
-    // 移除媒体占位符(U+FFFC)，避免发送给AI时产生干扰
-    final cleanContent = StringUtils.removeObjectReplacementChar(note.content);
-    // 重新生成时完全随机，跳过内容匹配
-    if (isRegeneration) {
-      final random = DateTime.now().millisecondsSinceEpoch % 3;
-      switch (random) {
-        case 0:
-          return AICardPrompts.randomStylePosterPrompt(
-            brandName: brandName,
-            content: cleanContent,
-            author: note.sourceAuthor,
-            date: _formatDate(note.date, languageCode: languageCode),
-            location: note.location,
-            weather: note.weather,
-            temperature: note.temperature,
-            dayPeriod: note.dayPeriod,
-            source: note.fullSource,
-          );
-        case 1:
-          return AICardPrompts.intelligentCardPrompt(
-            brandName: brandName,
-            content: cleanContent,
-            author: note.sourceAuthor,
-            date: _formatDate(note.date, languageCode: languageCode),
-            location: note.location,
-            weather: note.weather,
-            temperature: note.temperature,
-            dayPeriod: note.dayPeriod,
-            source: note.fullSource,
-          );
-        case 2:
-        default:
-          return AICardPrompts.contentAwareVisualPrompt(
-            brandName: brandName,
-            content: cleanContent,
-            author: note.sourceAuthor,
-            date: _formatDate(note.date, languageCode: languageCode),
-            location: note.location,
-            weather: note.weather,
-            temperature: note.temperature,
-            dayPeriod: note.dayPeriod,
-            source: note.fullSource,
-          );
-      }
-    }
-
-    // 分析内容特征
-    final content = cleanContent.toLowerCase();
-    final hasAuthor =
-        note.sourceAuthor != null && note.sourceAuthor!.isNotEmpty;
-
-    // 如果指定了自定义风格，使用对应的提示词
-    if (customStyle != null) {
-      switch (customStyle) {
-        case 'creative':
-          return AICardPrompts.randomStylePosterPrompt(
-            brandName: brandName,
-            content: cleanContent,
-            author: note.sourceAuthor,
-            date: _formatDate(note.date, languageCode: languageCode),
-            location: note.location,
-            weather: note.weather,
-            temperature: note.temperature,
-            dayPeriod: note.dayPeriod,
-            source: note.fullSource,
-          );
-        case 'intelligent':
-          return AICardPrompts.intelligentCardPrompt(
-            brandName: brandName,
-            content: cleanContent,
-            author: note.sourceAuthor,
-            date: _formatDate(note.date, languageCode: languageCode),
-            location: note.location,
-            weather: note.weather,
-            temperature: note.temperature,
-            dayPeriod: note.dayPeriod,
-            source: note.fullSource,
-          );
-        case 'visual':
-          return AICardPrompts.contentAwareVisualPrompt(
-            brandName: brandName,
-            content: cleanContent,
-            author: note.sourceAuthor,
-            date: _formatDate(note.date, languageCode: languageCode),
-            location: note.location,
-            weather: note.weather,
-            temperature: note.temperature,
-            dayPeriod: note.dayPeriod,
-            source: note.fullSource,
-          );
-      }
-    }
-
-    // 使用随机化的智能选择，避免单调
-    final random = DateTime.now().millisecondsSinceEpoch % 100;
-
-    // 1. 检查是否为引用/名言（30%概率使用智能卡片，70%随机海报）
-    if (hasAuthor ||
-        content.contains('说') ||
-        content.contains('曰') ||
-        content.contains('"')) {
-      if (random < 30) {
-        return AICardPrompts.intelligentCardPrompt(
-          brandName: brandName,
-          content: cleanContent,
-          author: note.sourceAuthor,
-          date: _formatDate(note.date, languageCode: languageCode),
-          location: note.location,
-          weather: note.weather,
-          temperature: note.temperature,
-          dayPeriod: note.dayPeriod,
-          source: note.fullSource,
-        );
-      } else {
-        return AICardPrompts.randomStylePosterPrompt(
-          brandName: brandName,
-          content: cleanContent,
-          author: note.sourceAuthor,
-          date: _formatDate(note.date, languageCode: languageCode),
-          location: note.location,
-          weather: note.weather,
-          temperature: note.temperature,
-          dayPeriod: note.dayPeriod,
-          source: note.fullSource,
-        );
-      }
-    }
-
-    // 2. 检查是否为技术/学习内容（40%视觉增强，30%智能，30%随机）
-    final techKeywords = ['代码', '编程', '算法', '技术', '开发', '学习', '知识', '方法', '原理'];
-    if (techKeywords.any((keyword) => content.contains(keyword))) {
-      if (random < 40) {
-        return AICardPrompts.contentAwareVisualPrompt(
-          brandName: brandName,
-          content: cleanContent,
-          author: note.sourceAuthor,
-          date: _formatDate(note.date, languageCode: languageCode),
-          location: note.location,
-          weather: note.weather,
-          temperature: note.temperature,
-          dayPeriod: note.dayPeriod,
-          source: note.fullSource,
-        );
-      } else if (random < 70) {
-        return AICardPrompts.intelligentCardPrompt(
-          brandName: brandName,
-          content: cleanContent,
-          author: note.sourceAuthor,
-          date: _formatDate(note.date, languageCode: languageCode),
-          location: note.location,
-          weather: note.weather,
-          temperature: note.temperature,
-          dayPeriod: note.dayPeriod,
-          source: note.fullSource,
-        );
-      } else {
-        return AICardPrompts.randomStylePosterPrompt(
-          brandName: brandName,
-          content: cleanContent,
-          author: note.sourceAuthor,
-          date: _formatDate(note.date, languageCode: languageCode),
-          location: note.location,
-          weather: note.weather,
-          temperature: note.temperature,
-          dayPeriod: note.dayPeriod,
-          source: note.fullSource,
-        );
-      }
-    }
-
-    // 3. 检查是否为情感/生活内容（50%随机海报，50%视觉增强）
-    final emotionalKeywords = ['感受', '心情', '生活', '感悟', '体验', '回忆', '梦想', '希望'];
-    if (emotionalKeywords.any((keyword) => content.contains(keyword))) {
-      if (random < 50) {
-        return AICardPrompts.randomStylePosterPrompt(
-          brandName: brandName,
-          content: cleanContent,
-          author: note.sourceAuthor,
-          date: _formatDate(note.date, languageCode: languageCode),
-          location: note.location,
-          weather: note.weather,
-          temperature: note.temperature,
-          dayPeriod: note.dayPeriod,
-          source: note.fullSource,
-        );
-      } else {
-        return AICardPrompts.contentAwareVisualPrompt(
-          brandName: brandName,
-          content: cleanContent,
-          author: note.sourceAuthor,
-          date: _formatDate(note.date, languageCode: languageCode),
-          location: note.location,
-          weather: note.weather,
-          temperature: note.temperature,
-          dayPeriod: note.dayPeriod,
-          source: note.fullSource,
-        );
-      }
-    }
-
-    // 4. 根据内容长度和随机性选择
-    if (cleanContent.length > 100) {
-      // 长内容：40%随机海报，30%智能，30%视觉
-      if (random < 40) {
-        return AICardPrompts.randomStylePosterPrompt(
-          brandName: brandName,
-          content: cleanContent,
-          author: note.sourceAuthor,
-          date: _formatDate(note.date, languageCode: languageCode),
-          location: note.location,
-          weather: note.weather,
-          temperature: note.temperature,
-          dayPeriod: note.dayPeriod,
-          source: note.fullSource,
-        );
-      } else if (random < 70) {
-        return AICardPrompts.intelligentCardPrompt(
-          brandName: brandName,
-          content: cleanContent,
-          author: note.sourceAuthor,
-          date: _formatDate(note.date, languageCode: languageCode),
-          location: note.location,
-          weather: note.weather,
-          temperature: note.temperature,
-          dayPeriod: note.dayPeriod,
-          source: note.fullSource,
-        );
-      } else {
-        return AICardPrompts.contentAwareVisualPrompt(
-          brandName: brandName,
-          content: cleanContent,
-          author: note.sourceAuthor,
-          date: _formatDate(note.date, languageCode: languageCode),
-          location: note.location,
-          weather: note.weather,
-          temperature: note.temperature,
-          dayPeriod: note.dayPeriod,
-          source: note.fullSource,
-        );
-      }
-    }
-
-    // 5. 默认使用三种提示词随机选择（各33%）
-    if (random < 33) {
-      return AICardPrompts.randomStylePosterPrompt(
-        brandName: brandName,
-        content: cleanContent,
-        author: note.sourceAuthor,
-        date: _formatDate(note.date, languageCode: languageCode),
-        location: note.location,
-        weather: note.weather,
-        temperature: note.temperature,
-        dayPeriod: note.dayPeriod,
-        source: note.fullSource,
-      );
-    } else if (random < 66) {
-      return AICardPrompts.intelligentCardPrompt(
-        brandName: brandName,
-        content: cleanContent,
-        author: note.sourceAuthor,
-        date: _formatDate(note.date, languageCode: languageCode),
-        location: note.location,
-        weather: note.weather,
-        temperature: note.temperature,
-        dayPeriod: note.dayPeriod,
-        source: note.fullSource,
-      );
-    } else {
-      return AICardPrompts.contentAwareVisualPrompt(
-        brandName: brandName,
-        content: cleanContent,
-        author: note.sourceAuthor,
-        date: _formatDate(note.date, languageCode: languageCode),
-        location: note.location,
-        weather: note.weather,
-        temperature: note.temperature,
-        dayPeriod: note.dayPeriod,
-        source: note.fullSource,
-      );
-    }
-  }
-
-  /// 静态元数据补全 (用于Isolate)
-  static String _ensureMetadataPresenceStatic(
-    String svg, {
-    required String brandName,
-    required String? date,
-    String? location,
-    String? weather,
-    String? temperature,
-    String? author,
-    String? source,
-    String? dayPeriod,
-    String languageCode = 'zh',
-    List<AICardProcessingLog>? logs,
-  }) {
-    final lower = svg.toLowerCase();
-    final hasDate = date != null && lower.contains(date.toLowerCase());
-    final hasLocation =
-        location != null && lower.contains(location.toLowerCase());
-    final hasWeather = weather != null && lower.contains(weather.toLowerCase());
-    final need = !(hasDate || hasLocation || hasWeather);
-    if (!need) {
-      return svg; // 已有至少一个信息
-    }
-    // 简单插入在 </svg> 前
-    final metaParts = <String>[];
-    // 规则：程序自动补全 -> 根据语言本地化
-    final localizedWeather =
-        _localizeWeather(weather, languageCode: languageCode);
-    final localizedDayPeriod =
-        _localizeDayPeriod(dayPeriod, languageCode: languageCode);
-
-    if (date != null) metaParts.add(date); // 已是格式化的
-    if (location != null) metaParts.add(location); // 用户输入不改动
-    if (localizedWeather != null) {
-      metaParts.add(
-        temperature != null
-            ? '$localizedWeather $temperature'
-            : localizedWeather,
-      );
-    }
-    if (author != null) metaParts.add(author);
-    if (source != null && source != author) metaParts.add(source);
-    if (localizedDayPeriod != null) metaParts.add(localizedDayPeriod);
-    metaParts.add(brandName);
-    final meta = metaParts.join(' · ');
-    final injection =
-        '<text x="200" y="590" text-anchor="middle" font-family="system-ui, -apple-system, sans-serif" font-size="10" fill="#ffffff" fill-opacity="0.75">${_escape(meta)}</text>';
-    final idx = svg.lastIndexOf('</svg>');
-    if (idx == -1) {
-      return svg; // 非法结构保持原样
-    }
-    return svg.substring(0, idx) + injection + svg.substring(idx);
-  }
-
-  static String _escape(String v) => v
-      .replaceAll('&', '&amp;')
-      .replaceAll('<', '&lt;')
-      .replaceAll('>', '&gt;');
-
-  // 检测是否包含中文
-  // ignore: unused_element
-  static bool _containsChinese(String text) =>
-      RegExp(r'[\u4e00-\u9fff]').hasMatch(text);
-
-  // 天气本地化映射 - 多语言支持
-  // Key: 英文标准名（小写）, Value: Map<languageCode, localizedName>
-  static const Map<String, Map<String, String>> _weatherLocalizations = {
-    // 基础天气类型
-    'clear': {'zh': '晴', 'en': 'Clear', 'ja': '晴れ', 'fr': 'Dégagé'},
-    'sunny': {'zh': '晴', 'en': 'Sunny', 'ja': '晴れ', 'fr': 'Ensoleillé'},
-    'partly_cloudy': {
-      'zh': '少云',
-      'en': 'Partly Cloudy',
-      'ja': '晴れ時々曇り',
-      'fr': 'Partiellement nuageux'
-    },
-    'cloudy': {'zh': '多云', 'en': 'Cloudy', 'ja': '曇り', 'fr': 'Nuageux'},
-    'overcast': {'zh': '阴', 'en': 'Overcast', 'ja': '曇天', 'fr': 'Couvert'},
-    'fog': {'zh': '雾', 'en': 'Fog', 'ja': '霧', 'fr': 'Brouillard'},
-    'haze': {'zh': '霾', 'en': 'Haze', 'ja': '霞', 'fr': 'Brume'},
-    'windy': {'zh': '有风', 'en': 'Windy', 'ja': '風', 'fr': 'Venteux'},
-
-    // 雨类天气
-    'drizzle': {'zh': '毛毛雨', 'en': 'Drizzle', 'ja': '霧雨', 'fr': 'Bruine'},
-    'light rain': {
-      'zh': '小雨',
-      'en': 'Light Rain',
-      'ja': '小雨',
-      'fr': 'Pluie légère'
-    },
-    'rain': {'zh': '雨', 'en': 'Rain', 'ja': '雨', 'fr': 'Pluie'},
-    'moderate rain': {
-      'zh': '中雨',
-      'en': 'Moderate Rain',
-      'ja': '中雨',
-      'fr': 'Pluie modérée'
-    },
-    'heavy rain': {
-      'zh': '大雨',
-      'en': 'Heavy Rain',
-      'ja': '大雨',
-      'fr': 'Forte pluie'
-    },
-    'freezing_rain': {
-      'zh': '冻雨',
-      'en': 'Freezing Rain',
-      'ja': '凍雨',
-      'fr': 'Pluie verglaçante'
-    },
-    'rain_shower': {
-      'zh': '阵雨',
-      'en': 'Rain Shower',
-      'ja': 'にわか雨',
-      'fr': 'Averse'
-    },
-    'thunderstorm': {
-      'zh': '雷雨',
-      'en': 'Thunderstorm',
-      'ja': '雷雨',
-      'fr': 'Orage'
-    },
-    'thunderstorm_heavy': {
-      'zh': '雷暴雨',
-      'en': 'Heavy Thunderstorm',
-      'ja': '激しい雷雨',
-      'fr': 'Orage violent'
-    },
-
-    // 雪类天气
-    'snow': {'zh': '雪', 'en': 'Snow', 'ja': '雪', 'fr': 'Neige'},
-    'light snow': {
-      'zh': '小雪',
-      'en': 'Light Snow',
-      'ja': '小雪',
-      'fr': 'Neige légère'
-    },
-    'heavy snow': {
-      'zh': '大雪',
-      'en': 'Heavy Snow',
-      'ja': '大雪',
-      'fr': 'Forte neige'
-    },
-    'snow_grains': {
-      'zh': '雪粒',
-      'en': 'Snow Grains',
-      'ja': '雪あられ',
-      'fr': 'Grésil'
-    },
-    'snow_shower': {
-      'zh': '阵雪',
-      'en': 'Snow Shower',
-      'ja': 'にわか雪',
-      'fr': 'Averse de neige'
-    },
-    'sleet': {'zh': '雨夹雪', 'en': 'Sleet', 'ja': 'みぞれ', 'fr': 'Grésil'},
-
-    // 其他格式变体
-    'rainy': {'zh': '雨', 'en': 'Rainy', 'ja': '雨', 'fr': 'Pluvieux'},
-    'snowy': {'zh': '雪', 'en': 'Snowy', 'ja': '雪', 'fr': 'Neigeux'},
-    'stormy': {'zh': '暴风雨', 'en': 'Stormy', 'ja': '嵐', 'fr': 'Orageux'},
-  };
-
-  /// 本地化天气字符串
-  /// [languageCode] 语言代码 ('zh', 'en', 'ja', 'fr', etc.)
-  static String? _localizeWeather(String? weather,
-      {String languageCode = 'zh'}) {
-    if (weather == null || weather.trim().isEmpty) return null;
-    final w = weather.toLowerCase().trim();
-
-    // 如果已经是目标语言，直接返回
-    if (_weatherLocalizations.containsKey(w)) {
-      return _weatherLocalizations[w]?[languageCode] ??
-          _weatherLocalizations[w]?['en'] ??
-          weather;
-    }
-
-    // 尝试匹配（可能输入带下划线或空格变体）
-    final normalized = w.replaceAll('_', ' ');
-    if (_weatherLocalizations.containsKey(normalized)) {
-      return _weatherLocalizations[normalized]?[languageCode] ??
-          _weatherLocalizations[normalized]?['en'] ??
-          weather;
-    }
-
-    // 未命中保持原样（可能已是目标语言）
-    return weather;
-  }
-
-  // 时间段本地化 - 多语言支持
-  static const Map<String, Map<String, String>> _dayPeriodLocalizations = {
-    'moring': {
-      'zh': '晨间',
-      'en': 'Morning',
-      'ja': '朝',
-      'fr': 'Matin'
-    }, // 兼容常见拼写错误
-    'morning': {'zh': '晨间', 'en': 'Morning', 'ja': '朝', 'fr': 'Matin'},
-    'noon': {'zh': '正午', 'en': 'Noon', 'ja': '正午', 'fr': 'Midi'},
-    'afternoon': {
-      'zh': '午后',
-      'en': 'Afternoon',
-      'ja': '午後',
-      'fr': 'Après-midi'
-    },
-    'evening': {'zh': '傍晚', 'en': 'Evening', 'ja': '夕方', 'fr': 'Soir'},
-    'night': {'zh': '夜晚', 'en': 'Night', 'ja': '夜', 'fr': 'Nuit'},
-    'dawn': {'zh': '黎明', 'en': 'Dawn', 'ja': '夜明け', 'fr': 'Aube'},
-    'dusk': {'zh': '黄昏', 'en': 'Dusk', 'ja': '夕暮れ', 'fr': 'Crépuscule'},
-    'late night': {
-      'zh': '深夜',
-      'en': 'Late Night',
-      'ja': '深夜',
-      'fr': 'Tard dans la nuit'
-    },
-    'midnight': {'zh': '午夜', 'en': 'Midnight', 'ja': '真夜中', 'fr': 'Minuit'},
-  };
-
-  /// 本地化时间段字符串
-  static String? _localizeDayPeriod(String? period,
-      {String languageCode = 'zh'}) {
-    if (period == null || period.trim().isEmpty) return null;
-    final p = period.toLowerCase().trim();
-
-    if (_dayPeriodLocalizations.containsKey(p)) {
-      return _dayPeriodLocalizations[p]?[languageCode] ??
-          _dayPeriodLocalizations[p]?['en'] ??
-          period;
-    }
-
-    // 未命中保持原样（可能已是目标语言）
-    return period;
-  }
-
-  /// 验证SVG基本结构
-  static bool _isValidSVGStructure(String svgContent) {
-    if (svgContent.trim().isEmpty) return false;
-
-    // 基本结构检查
-    if (!svgContent.contains('<svg') || !svgContent.contains('</svg>')) {
-      return false;
-    }
-
-    // 检查标签是否正确闭合（简单检查）
-    final openTags = '<svg'.allMatches(svgContent).length;
-    final closeTags = '</svg>'.allMatches(svgContent).length;
-    if (openTags != closeTags) {
-      return false;
-    }
-
-    // 检查是否有基本的SVG内容
-    if (svgContent.length < 50) {
-      // 太短的SVG可能无效
-      return false;
-    }
-
-    return true;
-  }
-
-  /// 标准化SVG属性
-  static String _normalizeSVGAttributes(String svgContent) {
-    String normalized = svgContent;
-
-    // 确保SVG有正确的命名空间
-    if (!normalized.contains('xmlns="http://www.w3.org/2000/svg"')) {
-      normalized = normalized.replaceFirst(
-        '<svg',
-        '<svg xmlns="http://www.w3.org/2000/svg"',
-      );
-    }
-
-    final inferredSize = _inferSvgIntrinsicSize(normalized);
-
-    // 移除现有的viewBox、width、height、preserveAspectRatio属性，避免AI返回的百分比尺寸导致错位
-    normalized = normalized
-        .replaceAll(RegExp(r'\s+viewBox="[^"]*"'), '')
-        .replaceAll(RegExp(r'\s+width="[^"]*"'), '')
-        .replaceAll(RegExp(r'\s+height="[^"]*"'), '')
-        .replaceAll(RegExp(r'\s+preserveAspectRatio="[^"]*"'), '');
-
-    // 统一设置标准属性: 仅设置viewBox与保留比例，不强制width/height，让预览与导出一致
-    normalized = normalized.replaceFirst(
-      '<svg',
-      '<svg viewBox="0 0 ${inferredSize.$1} ${inferredSize.$2}" preserveAspectRatio="xMidYMid meet"',
-    );
-
-    return normalized;
-  }
-
-  /// 推断SVG内在尺寸，忽略百分比/无效尺寸，防止viewBox被错误设置为100导致裁剪
-  static (String, String) _inferSvgIntrinsicSize(String svgContent) {
-    // 1) 优先使用合法的viewBox
-    final viewBoxMatch = _viewBoxRegex.firstMatch(svgContent);
-    if (viewBoxMatch != null) {
-      final parts = viewBoxMatch.group(1)!.split(_splitSpaceCommaRegex);
-      if (parts.length == 4 && parts.every((p) => double.tryParse(p) != null)) {
-        return (parts[2], parts[3]);
-      }
-    }
-
-    // 2) 解析根节点的width/height（忽略百分比）
-    double? w = _parseNumericDimension(
-      _widthRegex.firstMatch(svgContent)?.group(1),
-    );
-    double? h = _parseNumericDimension(
-      _heightRegex.firstMatch(svgContent)?.group(1),
-    );
-
-    // 3) 如果根尺寸不可靠，尝试从首个rect推断（通常为背景矩形）
-    if (w == null || h == null) {
-      final rectMatch = _rectFallbackRegex.firstMatch(svgContent);
-      if (rectMatch != null) {
-        w = _parseNumericDimension(rectMatch.group(1)) ?? w;
-        h = _parseNumericDimension(rectMatch.group(2)) ?? h;
-      }
-    }
-
-    // 4) 回退默认值
-    w ??= 400;
-    h ??= 600;
-
-    return (w.toString(), h.toString());
-  }
-
-  static double? _parseNumericDimension(String? raw) {
-    if (raw == null || raw.trim().isEmpty) return null;
-    if (raw.contains('%')) return null;
-    final cleaned = raw.replaceAll(_numericRegex, '');
-    if (cleaned.isEmpty) return null;
-    final parsed = double.tryParse(cleaned);
-    if (parsed == null || parsed <= 0) return null;
-    return parsed;
-  }
-
-  /// 验证SVG内容安全性
-  static bool _isSafeSVGContent(String svgContent,
-      [List<AICardProcessingLog>? logs]) {
-    // 检查是否包含潜在危险的元素
-    final dangerousElements = [
-      '<script',
-      '<iframe',
-      '<object',
-      '<embed',
-      'javascript:',
-      'data:text/html',
-      'onload=',
-      'onclick=',
-      'onerror=',
-    ];
-
-    final lowerContent = svgContent.toLowerCase();
-    for (final dangerous in dangerousElements) {
-      if (lowerContent.contains(dangerous)) {
-        final msg = '发现不安全的SVG元素: $dangerous';
-        if (logs != null) {
-          logs.add(AICardProcessingLog('WARN', msg));
-        } else {
-          AppLogger.w(msg, source: 'AICardGeneration');
-        }
-        return false;
-      }
-    }
-
-    return true;
-  }
-}
-
-/// DTO for passing data to the isolate
-///
-/// Contains all necessary information to process the SVG card in a background isolate.
-class AICardProcessingData {
-  final String svgContent;
-  final String brandName;
-  final String? date;
-  final String? location;
-  final String? weather;
-  final String? temperature;
-  final String? author;
-  final String? source;
-  final String? dayPeriod;
-  final String languageCode;
-
-  AICardProcessingData({
-    required this.svgContent,
-    required this.brandName,
-    required this.date,
-    this.location,
-    this.weather,
-    this.temperature,
-    this.author,
-    this.source,
-    this.dayPeriod,
-    this.languageCode = 'zh',
-  });
-}
-
-/// Log entry generated during isolate processing
-class AICardProcessingLog {
-  final String level; // 'DEBUG', 'INFO', 'WARN', 'ERROR'
-  final String message;
-
-  AICardProcessingLog(this.level, this.message);
-}
-
-/// DTO for returning results from the isolate
-class AICardProcessingResult {
-  final String svg;
-  final List<AICardProcessingLog> logs;
-
-  AICardProcessingResult({
-    required this.svg,
-    required this.logs,
-  });
 }
